Remove /modhelp (fixes #1401)
