+++ src/engine/client/updater.h
@@ -2,7 +2,7 @@
 #define ENGINE_CLIENT_UPDATER_H
 
 #include <engine/updater.h>
-#include <engine/shared/http.h>
+#include <engine/client/http.h>
 #include <map>
 #include <string>
