+++ src/game/server/ddracechat.cpp
@@ -1399,80 +1399,6 @@ void CGameContext::ConProtectedKill(IConsole::IResult *pResult, void *pUserData)
 	}
 }
 
-void CGameContext::ConModhelp(IConsole::IResult *pResult, void *pUserData)
-{
-	CGameContext *pSelf = (CGameContext *) pUserData;
-
-	if(!CheckClientID(pResult->m_ClientID))
-		return;
-
-	CPlayer *pPlayer = pSelf->m_apPlayers[pResult->m_ClientID];
-	if(!pPlayer)
-		return;
-
-	if(pPlayer->m_pPostJson)
-	{
-		pSelf->SendChatTarget(pResult->m_ClientID, "Your last request hasn't finished processing yet, please slow down.");
-		return;
-	}
-
-	int CurTick = pSelf->Server()->Tick();
-	if(pPlayer->m_ModhelpTick != -1)
-	{
-		int TickSpeed = pSelf->Server()->TickSpeed();
-		int NextModhelpTick = pPlayer->m_ModhelpTick + g_Config.m_SvModhelpDelay * TickSpeed;
-		if(NextModhelpTick > CurTick)
-		{
-			char aBuf[128];
-			str_format(aBuf, sizeof(aBuf), "You must wait %d seconds before you can execute this command again.",
-				(NextModhelpTick - CurTick) / TickSpeed);
-			pSelf->SendChatTarget(pResult->m_ClientID, aBuf);
-			return;
-		}
-	}
-
-	pPlayer->m_ModhelpTick = CurTick;
-
-	char aBuf[512];
-	str_format(aBuf, sizeof(aBuf), "Moderator help is requested by '%s' (ID: %d):",
-			pSelf->Server()->ClientName(pResult->m_ClientID),
-			pResult->m_ClientID);
-
-	// Send the request to all authed clients.
-	for(int i = 0; i < MAX_CLIENTS; i++)
-	{
-		if(pSelf->m_apPlayers[i] && pSelf->Server()->ClientAuthed(i))
-		{
-			pSelf->SendChatTarget(i, aBuf);
-			pSelf->SendChatTarget(i, pResult->GetString(0));
-		}
-	}
-	if(g_Config.m_SvModhelpUrl[0])
-	{
-		bool ModeratorPresent = false;
-		for(int i = 0; i < MAX_CLIENTS; i++)
-		{
-			if(pSelf->m_apPlayers[i] && pSelf->Server()->ClientAuthed(i))
-			{
-				ModeratorPresent = true;
-				break;
-			}
-		}
-
-		char aJson[512];
-		char aPlayerName[64];
-		char aMessage[128];
-		str_format(aJson, sizeof(aJson), "{\"port\":%d,\"moderator_present\":%s,\"player_id\":%d,\"blacklisted\":%s,\"player_name\":\"%s\",\"message\":\"%s\"}",
-			g_Config.m_SvPort,
-			JsonBool(ModeratorPresent),
-			pResult->m_ClientID,
-			!pSelf->Server()->DnsblWhite(pResult->m_ClientID) ? "true" : "false",
-			EscapeJson(aPlayerName, sizeof(aPlayerName), pSelf->Server()->ClientName(pResult->m_ClientID)),
-			EscapeJson(aMessage, sizeof(aMessage), pResult->GetString(0)));
-		pSelf->Engine()->AddJob(pPlayer->m_pPostJson = std::make_shared<CPostJson>(g_Config.m_SvModhelpUrl, false, aJson));
-	}
-}
-
 #if defined(CONF_SQL)
 void CGameContext::ConPoints(IConsole::IResult *pResult, void *pUserData)
 {
