+++ src/game/server/player.cpp
@@ -69,7 +69,6 @@ void CPlayer::Reset()
 	m_LastWhisperTo = -1;
 	m_LastSetSpectatorMode = 0;
 	m_TimeoutCode[0] = '\0';
-	m_ModhelpTick = -1;
 
 	m_TuneZone = 0;
 	m_TuneZoneOld = m_TuneZone;
