+++ src/game/server/player.h
@@ -6,7 +6,6 @@
 // this include should perhaps be removed
 #include "entities/character.h"
 #include "gamecontext.h"
-#include <engine/shared/http.h>
 
 // player object
 class CPlayer
@@ -175,7 +174,6 @@ public:
 	int m_ChatScore;
 
 	bool m_Moderating;
-	int m_ModhelpTick;
 
 	bool AfkTimer(int new_target_x, int new_target_y); //returns true if kicked
 	void AfkVoteTimer(CNetObj_PlayerInput *NewTarget);
@@ -198,7 +196,6 @@ public:
 #if defined(CONF_SQL)
 	int64 m_LastSQLQuery;
 #endif
-	std::shared_ptr<CPostJson> m_pPostJson;
 };
 
 #endif
